requirements.txt
# MCP Framework
fastmcp==2.0.0

# API Clients
google-analytics-data==0.21.0
requests==2.31.0

# AI Agent
anthropic==0.31.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Google OAuth
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Configuration
python-dotenv==1.0.1
pydantic==2.8.2

# Email
python-dateutil==2.8.2
secure-smtplib==0.1.1

# Utilities
httpx==0.27.0
aiohttp==3.11.0
pandas==2.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development
black==23.1.0
flake8==6.0.0
mypy==1.0.0
